Refactored check for preemptive authentication
- new protected method isPreemptiveAuthRequest() in AuthenticatorBase
which is overridden in some authenticators
- moved getRequestCertificates() from AuthenticatorBase to
SSLAuthenticator
